seperate profile hooks, contain home h1 display method, remove unused server files
