+++ client/src/App.tsx
@@ -7,21 +7,18 @@ import Profile from "./components/views/profile/Profile";
 import SearchLanding from "./components/views/searchLanding/SearchLanding";
 import { useSearch } from "./hooks/searchHook";
 import { useBlogData } from "./hooks/blogHook";
-import { useSinglePost } from "./hooks/postHook";
 import { useCreateUser } from "./hooks/createUserHook";
-import { useUpdateProfile, useGetProfile } from "./hooks/profileHook";
-import { useHome } from './hooks/homeDisplayHook';
+import { useGetProfile } from './hooks/getProfileHook'
+import { useUpdateProfile } from "./hooks/updateProfileHook";
 import Post from "./components/views/singlePost/post";
 import Create from "./components/views/create/create";
 
 export default function AllRoutes(children: any) {
     const { search, setSearch, results } = useSearch();
     const { category, setCategory } = useBlogData();
-    const { postData, setPostId } = useSinglePost();
     const { userExists, setUserExists, userId } = useCreateUser();
     const { formsOpen, setFormsOpen, setBio, setCreatures } = useUpdateProfile()
     const { profileData } = useGetProfile();
-    const { setHome } = useHome();
 
     return (
         <BrowserRouter>
@@ -32,8 +29,7 @@ export default function AllRoutes(children: any) {
                             userExists={userExists}
                             setUserExists={setUserExists} />}>
                         <Route path='/' element={<Homepage 
-                            setCategory={setCategory}
-                            setHome={setHome}/>} />
+                            setCategory={setCategory}/>} />
                         <Route path='/blog/:category' element={<Blog category={category}/>} />
                         <Route path='/dashboard' element={<Dashboard />} />
                         <Route path='/profile/:userId' element={<Profile 
@@ -46,7 +42,7 @@ export default function AllRoutes(children: any) {
                         <Route path='/search/:searchItems'element={<SearchLanding 
                         results={results} 
                         search={search} />} />
-                        <Route path='/post/:id' element={<Post postData={postData} setPostId={setPostId}/>} />
+                        <Route path='/post/:id' element={<Post/>} />
                     </Route>
                 </Routes>
         </BrowserRouter>
