+++ client/src/components/main/main.scss
@@ -43,29 +43,6 @@ h1.cryptid.homepage:hover    {
     color: black;
 }
 
-div.fonts   {
-    z-index: 10;
-    width: 100vw;
-    display:flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.title   {
-    font-family: var(--title-font);
-}
-
-.body    {
-    font-family: var(--body-font);
-}
-
-.sketchy   {
-    font-family: var(--special-font);
-}
-
-.other  {
-    font-family: var(--other-font);
-}
 
 @media screen and (max-width: 800px) {
     main    {
